refactor(auction): event를 domain 하위로 이동
