+++ auction-service/src/main/java/com/bugzero/rarego/app/AuctionCreateBidUseCase.java
@@ -3,8 +3,8 @@ package com.bugzero.rarego.app;
 import com.bugzero.rarego.domain.Auction;
 import com.bugzero.rarego.domain.AuctionMember;
 import com.bugzero.rarego.domain.Bid;
-import com.bugzero.rarego.event.AuctionBidCreatedEvent;
-import com.bugzero.rarego.event.AuctionUpdatedEvent;
+import com.bugzero.rarego.domain.event.AuctionBidCreatedEvent;
+import com.bugzero.rarego.domain.event.AuctionUpdatedEvent;
 import com.bugzero.rarego.global.exception.CustomException;
 import com.bugzero.rarego.global.response.ErrorType;
 import com.bugzero.rarego.in.dto.BidResponseDto;
