+++ auction-service/src/main/java/com/bugzero/rarego/app/AuctionSettlementSupport.java
@@ -3,12 +3,12 @@ package com.bugzero.rarego.app;
 import com.bugzero.rarego.domain.Auction;
 import com.bugzero.rarego.domain.AuctionOrder;
 import com.bugzero.rarego.domain.Bid;
-import com.bugzero.rarego.event.AuctionFailedEvent;
+import com.bugzero.rarego.domain.event.AuctionFailedEvent;
+import com.bugzero.rarego.global.exception.CustomException;
+import com.bugzero.rarego.global.response.ErrorType;
 import com.bugzero.rarego.out.AuctionOrderRepository;
 import com.bugzero.rarego.out.AuctionRepository;
 import com.bugzero.rarego.out.BidRepository;
-import com.bugzero.rarego.global.exception.CustomException;
-import com.bugzero.rarego.global.response.ErrorType;
 import com.bugzero.rarego.shared.auction.event.AuctionEndedEvent;
 import com.bugzero.rarego.shared.auction.type.AuctionStatus;
 import lombok.RequiredArgsConstructor;
