+++ auction-service/src/main/java/com/bugzero/rarego/domain/event/AuctionBidCreatedEvent.java
@@ -1,4 +1,4 @@
-package com.bugzero.rarego.event;
+package com.bugzero.rarego.domain.event;
 
 import java.time.LocalDateTime;
