+++ auction-service/src/test/java/com/bugzero/rarego/app/AuctionCreateBidUseCaseTest.java
@@ -3,11 +3,11 @@ package com.bugzero.rarego.app;
 import com.bugzero.rarego.domain.Auction;
 import com.bugzero.rarego.domain.AuctionMember;
 import com.bugzero.rarego.domain.Bid;
-import com.bugzero.rarego.event.AuctionBidCreatedEvent;
-import com.bugzero.rarego.in.dto.BidResponseDto;
-import com.bugzero.rarego.out.BidRepository;
+import com.bugzero.rarego.domain.event.AuctionBidCreatedEvent;
 import com.bugzero.rarego.global.exception.CustomException;
 import com.bugzero.rarego.global.response.ErrorType;
+import com.bugzero.rarego.in.dto.BidResponseDto;
+import com.bugzero.rarego.out.BidRepository;
 import com.bugzero.rarego.shared.auction.type.AuctionStatus;
 import com.bugzero.rarego.shared.payment.out.PaymentApiClient;
 import org.junit.jupiter.api.DisplayName;
